backend/src/services/rating.ts
import { Beach, BeachPosition } from "../models/beach";

// N, north: norte
// S, south, sul
// E, east, leste
// W, west, oeste

const waveHeights = {
  ankleToKnee: {
    min: 0.3,
    max: 1.0,
  },
  waistHigh: {
    min: 1.0,
    max: 2.0,
  },
  headHigh: {
    min: 2.0,
    max: 2.5,
  },
};

export class Rating {
  constructor(private beach: Beach) {}

  public getRatingBasedOnWindAndWavePositions(
    wavePosition: BeachPosition,
    windPosition: BeachPosition
  ): number {
    if (wavePosition === windPosition) {
      return 1;
    } else if (this.isWindOffShore(wavePosition, windPosition)) {
      return 5;
    }
    return 3;
  }

  public getRatingForSwellPeriod(period: number): number {
    if (period < 7) {
      return 1;
    }
    if (period < 10) {
      return 2;
    }
    if (period < 14) {
      return 4;
    }
    return 5;
  }

  public getRatingForSwellSize(height: number): number {
    if (height < waveHeights.ankleToKnee.min) {
      return 1;
    }
    if (height < waveHeights.ankleToKnee.max) {
      return 2;
    }
    if (height < waveHeights.waistHigh.max) {
      return 3;
    }
    return 5;
  }

  public getPositionFromLocation(coordinates: number): BeachPosition {
    if (coordinates < 50) return BeachPosition.N;
    if (coordinates < 120) return BeachPosition.E;
    if (coordinates < 220) return BeachPosition.S;
    if (coordinates < 310) return BeachPosition.W;
    return BeachPosition.N;
  }

  private isWindOffShore(
    wavePosition: BeachPosition,
    windPosition: BeachPosition
  ): boolean {
    return (
      (wavePosition === BeachPosition.N &&
        windPosition === BeachPosition.S &&
        this.beach.position === BeachPosition.N) ||
      (wavePosition === BeachPosition.S &&
        windPosition === BeachPosition.N &&
        this.beach.position === BeachPosition.S) ||
      (wavePosition === BeachPosition.E &&
        windPosition === BeachPosition.W &&
        this.beach.position === BeachPosition.E) ||
      (wavePosition === BeachPosition.W &&
        windPosition === BeachPosition.E &&
        this.beach.position === BeachPosition.W)
    );
  }
}
